# Optimize transpose performance