--- a/oneflow/user/kernels/transpose_kernel.cpp
+++ b/oneflow/user/kernels/transpose_kernel.cpp
@@ -25,17 +25,37 @@ template<DeviceType device_type, typename T>
 class TransposeKernel final : public OpKernel {
  public:
   TransposeKernel() = default;
-  ~TransposeKernel() = default;
+  ~TransposeKernel() override = default;
 
  private:
   void Compute(KernelComputeContext* ctx) const override {
     const Tensor* tensor_in = ctx->Tensor4ArgNameAndIndex("input", 0);
     Tensor* tensor_out = ctx->Tensor4ArgNameAndIndex("output", 0);
     const auto& perm = ctx->Attr<std::vector<int32_t>>("perm");
-    NewKernelUtil<device_type>::Transpose(ctx->device_ctx(), tensor_in->shape().NumAxes(),
-                                          tensor_in->shape(), tensor_out->shape(),
-                                          StdVec2PbRf(perm), tensor_in->shape().elem_cnt(),
-                                          tensor_in->dptr<T>(), tensor_out->mut_dptr<T>());
+    using PackType = int64_t;
+    const size_t num_elem_per_pack = sizeof(PackType) / sizeof(T);
+    const ShapeView& in_shape = tensor_in->shape();
+    const ShapeView& out_shape = tensor_out->shape();
+    if (num_elem_per_pack != 1 && perm.back() == perm.size() - 1
+        && in_shape.At(in_shape.NumAxes() - 1) % num_elem_per_pack == 0) {
+      CHECK_EQ(in_shape.At(in_shape.NumAxes() - 1), out_shape.At(out_shape.NumAxes() - 1));
+      DimVector packed_in_dim_vec;
+      in_shape.ToDimVector(&packed_in_dim_vec);
+      packed_in_dim_vec.back() /= num_elem_per_pack;
+      const Shape packed_in_shape(packed_in_dim_vec);
+      DimVector packed_out_dim_vec;
+      out_shape.ToDimVector(&packed_out_dim_vec);
+      packed_out_dim_vec.back() /= num_elem_per_pack;
+      const Shape packed_out_shape(packed_out_dim_vec);
+      NewKernelUtil<device_type>::Transpose(
+          ctx->device_ctx(), packed_in_shape.NumAxes(), packed_in_shape, packed_out_shape, perm,
+          packed_in_shape.elem_cnt(), reinterpret_cast<const PackType*>(tensor_in->dptr<T>()),
+          reinterpret_cast<PackType*>(tensor_out->mut_dptr<T>()));
+    } else {
+      NewKernelUtil<device_type>::Transpose(ctx->device_ctx(), in_shape.NumAxes(), in_shape,
+                                            tensor_out->shape(), perm, in_shape.elem_cnt(),
+                                            tensor_in->dptr<T>(), tensor_out->mut_dptr<T>());
+    }
   }
   bool AlwaysComputeWhenAllOutputsEmpty() const override { return false; }
 };