Benchmarks of bert base on RTX 2080ti, batch size is 64.

Old + FP32

```
                    2.21%  1.20590s      9600  125.61us  119.23us  212.55us  void oneflow::_GLOBAL__N__72_tmpxft_00009efe_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_62e2f439::TransposeGpu<int=4, float>(oneflow::_GLOBAL__N__72_tmpxft_00009efe_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_62e2f439::Int32Array<int=4>, oneflow::_GLOBAL__N__72_tmpxft_00009efe_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_62e2f439::Int32Array, int, float const *, oneflow::_GLOBAL__N__72_tmpxft_00009efe_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_62e2f439::Int32Array<int=4>*)

```

Old + FP16

```
                    4.50%  917.22ms      9600  95.543us  88.448us  101.54us  void oneflow::_GLOBAL__N__72_tmpxft_00009efe_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_62e2f439::TransposeGpu<int=4, __half>(oneflow::_GLOBAL__N__72_tmpxft_00009efe_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_62e2f439::Int32Array<int=4>, oneflow::_GLOBAL__N__72_tmpxft_00009efe_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_62e2f439::Int32Array, int, __half const *, oneflow::_GLOBAL__N__72_tmpxft_00009efe_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_62e2f439::Int32Array<int=4>*)

```

New + FP32

```
                    1.73%  940.97ms      9600  98.017us  96.352us  105.63us  void oneflow::_GLOBAL__N__72_tmpxft_00008b90_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_99915e63::TransposeGpu<int=4, long>(oneflow::_GLOBAL__N__72_tmpxft_00008b90_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_99915e63::Int32Array<int=4>, oneflow::_GLOBAL__N__72_tmpxft_00008b90_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_99915e63::Int32Array, int, long const *, oneflow::_GLOBAL__N__72_tmpxft_00008b90_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_99915e63::Int32Array<int=4>*)

```

New + FP16

```
                    2.44%  481.55ms      9600  50.161us  48.288us  52.768us  void oneflow::_GLOBAL__N__72_tmpxft_00008b90_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_99915e63::TransposeGpu<int=4, long>(oneflow::_GLOBAL__N__72_tmpxft_00008b90_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_99915e63::Int32Array<int=4>, oneflow::_GLOBAL__N__72_tmpxft_00008b90_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_99915e63::Int32Array, int, long const *, oneflow::_GLOBAL__N__72_tmpxft_00008b90_00000000_9_cuda_arithemetic_interface_compute_75_cpp1_ii_99915e63::Int32Array<int=4>*)

```