@chengtbf 2020-08-16
我们现在的cnn里不用transpose了。是crop mirror normalize，里面会自动做transpose。这个优化还有意义吗